Update the hook for extended Rich Text features and add body-text-large class
