+++ foundation_cms/legacy_apps/wagtailpages/wagtail_hooks.py
@@ -59,52 +59,6 @@ post_save.disconnect(sync_trees_on_locale_sync_save, sender=LocaleSynchronizatio
 post_save.connect(sync_view_restriction_change, sender=PageViewRestriction)
 
 
-# Extended rich text features for our site
-@hooks.register("register_rich_text_features")
-def register_large_feature(features):
-    """
-    Registering the 'large' Draftail feature which
-    adds a span around the selected text with its class
-    set to 'tw-body-large'
-    """
-
-    # 1. Set up variables for use below
-    feature_name = "large"
-    type_ = "LARGE"
-
-    # 2. Set up a dictionary to pass to Draftail to configure
-    # how it handles this feature in its toolbar.
-    # The 'style' attribute controls how Draftail formats that text
-    # in the editor - does not affect the final rendered HTML
-    # In this case, I am adding similar formatting to what
-    # the CSS will do to that 'small-caption' class in the template
-    control = {
-        "type": type_,
-        "label": "L",
-        "description": "Large body text",
-        "style": {"font-size": "125%"},
-    }
-
-    # 3. Call register_editor_plugin to register the configuration for Draftail.
-    features.register_editor_plugin("draftail", feature_name, draftail_features.InlineStyleFeature(control))
-
-    # 4.configure the content transform from the DB to the editor and back.
-
-    # The "From" version uses a CSS selector to find spans with a class of 'tw-body-large'
-    # The "To" version adds a span with a class of 'tw-body-large' surrounding the selected text
-    db_conversion = {
-        "from_database_format": {'span[class="tw-body-large"]': InlineStyleElementHandler(type_)},
-        "to_database_format": {"style_map": {type_: 'span class="tw-body-large"'}},
-    }
-
-    # 5. Call register_converter_rule to register the content transformation conversion.
-    features.register_converter_rule("contentstate", feature_name, db_conversion)
-
-    # 6. (optional) Add the feature to the default features list to make it available
-    # on rich text fields that do not specify an explicit 'features' list
-    features.default_features.append("large")
-
-
 # Updating external links in rich text blocks to open in a new tab
 class RichTextExternalLinkNewTabHandler(LinkHandler):
     identifier = "external"
